feat(HTD-51): add TagsNavigation
