+++ src/components/sections/blog/overview-explorer/index.tsx
@@ -1,45 +1,15 @@
-import {
-  StyledExplorerContainer,
-  StyledExplorerSection,
-  StyledSearchWrapper,
-  StyledTagText,
-  StyledTagsWrapper,
-} from './overview-explorer.styled';
-import { TagsDropwdown } from './tags-dropwdown';
+import { Suspense } from 'react';
 
-import { Link } from '@/components/elements/link';
-import { getAllCategoryTags } from '@/lib/mdx/blog';
+import { StyledExplorerContainer, StyledExplorerSection, StyledSearchWrapper } from './overview-explorer.styled';
+import { TagsNavigation } from './tags-navigation';
 
-const FEATURED_TAGS_COUNT = 5;
-
-export const OverviewExplorer = async () => {
-  const tags = await getAllCategoryTags();
-  // TODO: replace with choosen featured categories later
-  const featuredTags = tags.slice(0, FEATURED_TAGS_COUNT);
-  const ramainingTags = tags.slice(FEATURED_TAGS_COUNT, tags.length);
-
+export const OverviewExplorer = () => {
   return (
     <StyledExplorerSection>
       <StyledExplorerContainer>
-        <StyledTagsWrapper aria-label="Blog post tags">
-          <ul>
-            <li>
-              <Link href={`/blog`}>
-                <StyledTagText>All posts</StyledTagText>
-              </Link>
-            </li>
-            {featuredTags.map(tag => (
-              <li key={tag}>
-                <Link href={`/blog?tags=${tag}`}>
-                  <StyledTagText>{tag}</StyledTagText>
-                </Link>
-              </li>
-            ))}
-            <li>
-              <TagsDropwdown tags={ramainingTags} />
-            </li>
-          </ul>
-        </StyledTagsWrapper>
+        <Suspense fallback={'loading tags...'}>
+          <TagsNavigation />
+        </Suspense>
         <StyledSearchWrapper>Search box</StyledSearchWrapper>
       </StyledExplorerContainer>
     </StyledExplorerSection>
