+++ src/components/sections/blog/overview-explorer/tags-navigation.tsx
@@ -1,0 +1,36 @@
+import { StyledTagText, StyledTagsWrapper } from './overview-explorer.styled';
+import { TagsDropwdown } from './tags-dropwdown';
+
+import { Link } from '@/components/elements/link';
+import { getAllCategoryTags } from '@/lib/mdx/blog';
+
+const FEATURED_TAGS_COUNT = 5;
+
+export const TagsNavigation = async () => {
+  const tags = await getAllCategoryTags();
+  // TODO: replace with choosen featured categories later
+  const featuredTags = tags.slice(0, FEATURED_TAGS_COUNT);
+  const ramainingTags = tags.slice(FEATURED_TAGS_COUNT, tags.length);
+
+  return (
+    <StyledTagsWrapper aria-label="Blog post tags">
+      <ul>
+        <li>
+          <Link href={`/blog`}>
+            <StyledTagText>All posts</StyledTagText>
+          </Link>
+        </li>
+        {featuredTags.map(tag => (
+          <li key={tag}>
+            <Link href={`/blog?tags=${tag}`}>
+              <StyledTagText>{tag}</StyledTagText>
+            </Link>
+          </li>
+        ))}
+        <li>
+          <TagsDropwdown tags={ramainingTags} />
+        </li>
+      </ul>
+    </StyledTagsWrapper>
+  );
+};
